+++ controllers/userController.js
@@ -125,9 +125,3 @@ exports.detail = async function(params) {
   const user = await User.findById(userId).select('id displayName followingCount followerCount lastActive viewCount');
   return apiSuccess(user);
 };
-
-exports.poems = async function(params) {
-  const userId = getUserId(params.token);
-  const poems = await Poem.find({author: userId}).exec();
-  return apiSuccess(poems);
-};
