+++ routes/v1/usersRouter.js
@@ -79,14 +79,4 @@ router.post('/detail', createTypeChecker({
   }));
 });
 
-router.post('/poems', createTypeChecker({
-  'token': STRING,
-}), createTokenChecker(), async (req, res) => {
-  const token = req.body.token;
-
-  res.json(await userController.poems({
-    token,
-  }));
-});
-
 module.exports = router;
